ThreadFix 1.0 Beta 13
Deleting a WAF now correctly deletes its rules
Removing a WAF from an Application deletes the WAF rules
API Keys now have a restricted attribute that only allows read
operations
Merging correctly updates severities for vulnerabilities
Qualys vulnerabilities have associated severities

+++ src/main/java/com/denimgroup/threadfix/service/WafServiceImpl.java
@@ -89,7 +89,16 @@ public class WafServiceImpl implements WafService {
 	@Override
 	@Transactional(readOnly = false)
 	public void deleteById(int wafId) {
-		wafDao.deleteById(wafId);
+		Waf waf = loadWaf(wafId);
+		
+		if (waf != null) {
+			if (waf.getWafRules() != null) {
+				for (WafRule rule : waf.getWafRules()) {
+					wafRuleDao.delete(rule);
+				}
+			}
+			wafDao.deleteById(wafId);
+		}
 	}
 
 	@Override
